requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
billiard==4.2.1
boto3==1.34.119
botocore==1.34.119
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.3
cron-descriptor==1.4.5
cryptography==42.0.7
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1.7
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-ses==4.1.0
django-spaghetti-and-meatballs==0.4.2
django-storages==1.14.3
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.28.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==30.8.0
freezegun==1.5.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
s3transfer==0.10.1
simplejson==3.19.3
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.2
typing_extensions==4.12.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
